Made more robust. Original: fd30f5f5659d0583b20be57ae02626c74734174f

+++ lal/packages/pulsar/test/ephemtoilwd.sh
@@ -6,7 +6,7 @@
 
 # USAGE:
 #
-# ilwdem ephemeris.dat > ephemeris.ilwd
+# ephemtoilwd ephemeris.dat > ephemeris.ilwd
 
 # ASSUMPTIONS:
 #
@@ -26,8 +26,29 @@
 #   components of each) respectively.
 #   NOTE: each entry spans 4 lines
 
+infile=${1:-/dev/null/xxx}
+if test "$infile" = "/dev/null/xxx"
+then
+  echo "Usage: $0 ephem.dat" >&2
+  exit 1
+fi
 
-infile=$1
+fail(){
+echo "$0: error in processing file $infile" >&2
+exit 1
+}
+
+if test ! -r $infile
+then
+  if test -f $infile
+  then
+    echo "$0: can't read $1: Permission denied" >&2
+  else
+    echo "$0: can't read $1: No such file" >&2
+  fi
+  exit 1
+fi
+
 tmpfile=$infile.tmp
 rm -f $tmpfile
 
@@ -35,6 +56,10 @@ rm -f $tmpfile
 nlines=`sed -n '1s/[^0-9]*\([0-9][0-9]*\)[^0-9][^0-9]*\([0-9][0-9]*\)\.[0-9]*[^0-9][^0-9]*\([0-9][0-9]*\)/\3/p' $infile`
 tstart=`sed -n '1s/[^0-9]*\([0-9][0-9]*\)[^0-9][^0-9]*\([0-9][0-9]*\)\.[0-9]*[^0-9][^0-9]*\([0-9][0-9]*\)/\1/p' $infile`
 tstep=`sed -n '1s/[^0-9]*\([0-9][0-9]*\)[^0-9][^0-9]*\([0-9][0-9]*\)\.[0-9]*[^0-9][^0-9]*\([0-9][0-9]*\)/\2/p' $infile`
+
+test "${nlines:-xxx}" != "xxx" || fail
+test "${tstart:-xxx}" != "xxx" || fail
+test "${tstep:-xxx}" != "xxx" || fail
 
 # Rest of the file has spurious newlines... correct this!
 # Also get rid of multiple spaces and initial spaces.
@@ -49,15 +74,22 @@ s/\n/ /
 s/  */ /g
 s/^  *//
 p
-}' $infile > $tmpfile
+}' $infile > $tmpfile || fail
 
 # Re-read start and stop times
 tstart=`sed -n '1s/[^0-9].*$//p' $tmpfile`
 tstop=`sed -n '$s/[^0-9].*$//p' $tmpfile`
 
+test "${tstart:-xxx}" != "xxx" || fail
+test "${tstop:-xxx}" != "xxx" || fail
+
 # Sanity check
 dt1=`expr $nlines \* $tstep - $tstep`
 dt2=`expr $tstop - $tstart`
+
+test "${dt1:-xxx}" != "xxx" || fail
+test "${dt2:-xxx}" != "xxx" || fail
+
 if test $dt1 -ne $dt2
 then
   echo "sanity check failed: time step seems wrong"
@@ -79,7 +111,7 @@ printf "    <int_4u name='gps_sec:stop_time' units='nanosec'>0</int_4u>\n"
 printf "    <real_4 name='time:step_size' units='sec'>$tstep</real_4>\n"
 printf "    <real_4 dims='9,$nlines' name='data' ndim='2' units='s,s,s,s/s,s/s,s/s,1/s,1/s,1/s'>"
 # The data: delete GPS time from data file and remove newlines.
-sed 's/[^ ]* //' $tmpfile | tr '\n' ' '
+sed 's/[^ ]* //' $tmpfile | tr '\n' ' ' || fail
 printf "</real_4>\n"
 printf "  </ilwd>\n"
